Do not release on every push to master
It doesn't make sense to do a release to PyPI for every commit
pushed to the master. We should instead do the releases when
we are sure that the code in the master branch is OK to release.

Hence, this PR makes the release workflow run on the `workflow_dispatch`
event, which is a manual trigger.