Do not expose the fields of the clients (#28)
We should keep the public API surface as minimal as possible.
I don't see the benefit of making the configuration related
fields of the client public. Hence, I converted all of them
to protected fields.

I have also updated the README in couple of places. I have removed
the part where we mention you could have per-command-formatting
configuration: It is a pretty bad idea to allow mutable state
on the client object.