Improve session management of the clients (#29)
* Improve session management of the clients

We had an implementation of the AsyncRedis that is not usable
without the `async with` block. Not all users are going to
use it that way, we should allow usages without it as well.

I also found the way the tests written not ideal. We had a
single sync and async client and reuse that throughout the
test suite. I believe the proper way is to use fixtures
per tests and make sure that the clients are closed
no matter the result of the test. That resulted in a huge
change, but I believe it was worth it. I believe the tests
are a bit better right now, as we now have proper test
isolation.

Also, the close method was missing from the AsyncRedis,
so I added it as well.

* replace code that leaks the connection