Use a string value for the scan cursor
Since it can be quite big, sending it as a number can cause it
to be larger than the max safe integer value in JSON, and result
in unpredictable iteration.

Instead, it is now converted to string.

We don't have to do it for other scan commands, their cursor values
are small, but I have done it anyways to be on the safe side.